try deploy

+++ middleware/src/c_utils.c
@@ -26,7 +26,7 @@ unsigned char reverse_bits(unsigned char b)
 
 float linear_interpolate(float x, float x1, float x2, float y1, float y2)
 {
-	assert(fabs(x2 - x1) > FLOAT_EPSILON);
+	assert(fabsf(x2 - x1) > FLOAT_EPSILON);
 
 	return y1 + ((x - x1) * (y2 - y1) / (x2 - x1));
 }
